Neural Network
A neural network consists of layers. The first layer is the input and the last layer is the output. The layers between the input and output are called hidden layers. The nodes of one layer can be connected to the nodes of the next layer.  Each connection between two nodes (called an edge) has a weight. 

Write a java code which  does the following :

 It takes the number of layers, and nodes in each layer as input, and creates a class object.
 It sets each edge's weight from user input
It should be possible to query the weight of any edge, given the node. 

Your program should  follow the principles of object oriented program